[js] rename showPdf.js -> pdf.js
